refactored controller to retun objects

+++ src/main/java/com/glen/todoapp/todo/ToDoRepository.java
@@ -1,9 +1,12 @@
 package com.glen.todoapp.todo;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ToDoRepository extends JpaRepository<ToDo, Integer>{
+public interface ToDoRepository extends JpaRepository<ToDo, Long>{
+
+	
 
 }
